Fixed #4138 - (de)serialization of JSON number types should keep the value as a double, instead of decimal, to avoid number to string conversion

+++ src/MassTransit/Serialization/JsonConverters/CaseInsensitiveDictionaryStringObjectJsonConverter.cs
@@ -195,7 +195,7 @@ namespace MassTransit.Serialization.JsonConverters
                     if (reader.TryGetInt64(out var result))
                         return result;
 
-                    return reader.GetDecimal();
+                    return reader.GetDouble();
 
                 case JsonTokenType.StartObject:
                     return ReadObject(ref reader, options);
